+++ TetrisWebAssembly/Models/TetrisBlock.cs
@@ -2,11 +2,11 @@
 
 public class TetrisBlock
 {
-    public int X { get; set; }
-    public int Y { get; set; }
+    public double X { get; set; }
+    public double Y { get; set; }
     public string Color { get; set; } // Store the color of the block
 
-    public TetrisBlock(int x, int y, string color)
+    public TetrisBlock(double x, double y, string color)
     {
         X = x;
         Y = y;
